Add three new events for commanded state
of azimuth and the two doors.
I considered combining these with the existing state events,
but the commanded state is affected by the command received by the CSC,
whereas the state is affected by what the TCP/IP interface reports via polling.
These will not be well synchronized so the output would likely be confusing
when a move started.