MTMount telemetry: complete overhaul
The new data is more complete and comes from parsing
the TMA's telemetry configuration file so the order
of some fields has changed in existing topics.

Also modify test_Units.py to allow mmWC and move the allowed
nonstandard units to a new constant, so the list is easier
to find and maintain.